BST-12156: add keyscope post-processor to gitleaks
Adds a new post-processor to the existing gitleaks scanner to run
keyscope on the gitleaks results to check if any found secret is still
active or not.

The gitleaks post-processor was also update to return new properties on
the sarif results that maps to a secret finding type. Also, it no longer
strip the secret values, since keyscope needs them to work. This was not
an issue, because the cli sarif processor would always remove them
anyway.