BST-13663 Move the rules from baseline to scanner
The baseline scanner was created to be included in the boost-scanner,
allowing order and streamline of definition.  The side effect of this is
that the rules of the baseline scanner show up in the policy, under the
same name a boost-scanner.

Updating the CLI to support moving the steps definition in the
boost-scanner is not trivial, so for now we'll just remove the rules
from the baseline and declare them directly in scanner.  This makes the
baseline not really usable on its own, as it would drop all the rules,
but it was never expect to be.