BST-17950 Revert adding secret scanning to trivy-fs
The story that added this was aimed only at updating trivy-image since
that's the only secret scanning we currently have for images.

For source-code, without proper benchmarking, we don't want to offer
trivy secret scanning as an alternative to gitleaks.  Maybe it will come
but not for now.

This change will prevent the trivy-fs scans from bearing the "secrets"
scan-type which shows up in the secret section of the scanner coverage.